src/test/kotlin/Day9Test.kt
import org.junit.jupiter.api.AfterEach
import org.junit.jupiter.api.Assertions.assertEquals
import org.junit.jupiter.api.BeforeEach
import org.junit.jupiter.api.Test
import org.junit.jupiter.api.TestInstance
import org.junit.jupiter.params.ParameterizedTest
import org.junit.jupiter.params.provider.MethodSource
import java.util.*

@TestInstance(TestInstance.Lifecycle.PER_CLASS)
internal class Day9Test {

    private val input = StandardInputStream()
    private val output = StandardOutputStream()

    @BeforeEach
    fun setUp() {
        System.setIn(input)
        System.setOut(output)
    }

    @AfterEach
    fun tearDown() {
        System.setIn(null)
        System.setOut(null)
    }

    @Test
    fun solve1() {
        val input = """
            R 4
            U 4
            L 3
            D 1
            R 4
            D 1
            L 5
            R 2
        """.trimIndent().split("\n").map { it.split(" ").zipWithNext { a, b -> Pair(a, b.toInt()) }.single() }
        val obj = Day9()
        assertEquals(13, obj.solve1(input))
    }

    @Test
    fun solve2() {
        val input = """
            R 4
            U 4
            L 3
            D 1
            R 4
            D 1
            L 5
            R 2
        """.trimIndent().split("\n").map { it.split(" ").zipWithNext { a, b -> Pair(a, b.toInt()) }.single() }
        val obj = Day9()
        assertEquals(1, obj.solve2(input))
    }
}
